pyproject.toml
[project]
name = "go-kangnam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.4",
    "pandas>=3.0.1",
    "publicdatareader>=1.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
]
